+++ app/src/main/java/com/rodcollab/cliq/collections/ClientItem.kt
@@ -1,0 +1,8 @@
+package com.rodcollab.cliq.collections
+
+data class ClientItem(
+    val id: String,
+    val name: String,
+    val reference: String,
+    val address: String
+)
